requirements.txt
apache-airflow
dbt-duckdb
pandas
requests
streamlit
plotly
numpy
SQLAlchemy>=1.4 